requirements.txt
Django==5.0
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.11.0
Pillow==10.1.0
openai==1.3.0
python-dotenv==1.0.0
